requirements.txt
sklearn
numpy
pandas
python-mnist